# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SaaS Boilerplate API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    delete:
      operationId: deleteMe
      tags: [users]
      summary: Soft delete current user (GDPR)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account scheduled for deletion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/sessions:
    get:
      operationId: getSessions
      tags: [users]
      summary: Get active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

  # PAYMENTS
  /payments/checkout:
    post:
      operationId: createCheckoutSession
      tags: [payments]
      summary: Create Stripe checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /payments/portal:
    get:
      operationId: getPortalUrl
      tags: [payments]
      summary: Get Stripe customer portal URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"
  /payments/invoices:
    get:
      operationId: getInvoices
      tags: [payments]
      summary: Get invoice history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
  /payments/subscription:
    get:
      operationId: getSubscription
      tags: [payments]
      summary: Get current subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, USER]
        plan:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
        emailVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        banned:
          type: boolean
      required:
        - id
        - email
        - name
        - role
        - plan
        - emailVerified
        - twoFactorEnabled
        - createdAt
        - banned
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string
    ChangePasswordBody:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - oldPassword
        - newPassword
    Session:
      type: object
      properties:
        id:
          type: string
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt
        - expiresAt
    CheckoutBody:
      type: object
      properties:
        plan:
          type: string
          enum: [PRO, ENTERPRISE]
      required:
        - plan
    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    PortalResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    Subscription:
      type: object
      properties:
        id:
          type: string
        plan:
          type: string
          enum: [FREE, PRO, ENTERPRISE]
        status:
          type: string
        currentPeriodStart:
          type: string
          format: date-time
          nullable: true
        currentPeriodEnd:
          type: string
          format: date-time
          nullable: true
        cancelAt:
          type: string
          format: date-time
          nullable: true
      required:
        - plan
        - status
    InvoiceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    Invoice:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        pdfUrl:
          type: string
          nullable: true
      required:
        - id
        - amount
        - currency
        - status
        - createdAt
...